+++ src/main/java/com/bitcrunzh/generic/config/reflection/annotation/property/ShortProperty.java
@@ -7,26 +7,26 @@ import java.lang.annotation.Target;
 
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.FIELD)
-public @interface IntProperty {
+public @interface ShortProperty {
     /**
      * @return the minimum (inclusive) valid value.
      */
-    int minValue() default Integer.MIN_VALUE;
+    short minValue() default Short.MIN_VALUE;
     /**
      * @return the maximum (inclusive) valid value.
      */
-    int maxValue() default Integer.MAX_VALUE;
+    short maxValue() default Short.MAX_VALUE;
 
     /**
      * White list of valid values.
      * An empty array means all values.
      */
-    int[] validValues() default {};
+    short[] validValues() default {};
 
     /**
      * Default recommended value for this property.
      */
-    int defaultValue() default 0;
+    short defaultValue() default 0;
 
     /**
      * Name of the property. If empty, the field name will be used.
